Reformat optimization code and apply best practise for pyhton
