+++ Constants.cs
@@ -9,6 +9,6 @@ public static class Constants
     public const string BaseUri = "https://www.sec.gov/";
     public const string CompaniesApi = "files/company_tickers_exchange.json";
     public const string CompanyFactsApi = "Archives/edgar/daily-index/xbrl/companyfacts.zip";
-    public static string OptimizationScriptPath =
-        Path.Combine(AppDomain.CurrentDomain.BaseDirectory, @"..\..\..\Optimization", "find_best_company.py");
+    public static readonly string OptimizationScriptPath =
+        Path.Combine(AppDomain.CurrentDomain.BaseDirectory, @"..\..\..\Optimization", "main.py");
 }
