+++ Optimization/requirements.txt
@@ -1,0 +1,3 @@
+pulp
+yfinance
+matplotlib
